Separate the LegacySourceService connection
Simplify LegacySourceManager by splitting the LegacySourceService connection logic into a separate class.
